Add files via upload
Without a navy campaign (something else I'm tinkering with) it's a little difficult to balance so currently once you unlock the first license just play through the story for the rest.

Possible reasons to not be seeing the licenses...

There is a reputation trigger with the republic to not be hostile for this to implement the first license unlock. You must also have received news about the secession.
This should trigger before you can officially join the free worlds but i did put it in as a requirement to not have joined because who would sell ships to their enemies?
So, it may conflict with preexisting saves.

At the current version I have only tested up until the first License unlock but will be actively working on this and other mods for some time yet.